hackathon_scraper/requirements.txt
requests==2.32.4
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11
pyyaml==6.0.2
python-dotenv==1.0.1
tqdm==4.66.5
tenacity==9.0.0
pandas==2.2.2
openpyxl==3.1.5
pydantic==2.11.7
python-dateutil==2.9.0.post0
